"Fully qualify" member names.
A concern with *just* using the interface name is when the same
interface name exists in multiple packages, and you're implementing
*all* of them:

    package p1;
    public interface List {}

    package p2;
    public interface List {}

    package example;
    public interface MyList implements p1.List, p2.List {}

With this API, the interface invoker for `IMyListInvoker` would
try to emit `_members_List` *twice*, which is a good way to get
compilation errors.

To avoid this, we could *hash* the package name + type name, as we
do with Java Callable Wrappers, but if you see `_members_hash1` vs.
`_members_hash2`, it's difficult to review what's going on.

Instead, use the full Java-side name, "mangled" so that it can be
an identifier.  This results in:

    partial class IMyListInvoker : Java.Lang.Object, IMyList {
        static readonly JniPeerMembers _members_p1_List = …;
        static readonly JniPeerMembers _members_p2_List = …;
        static readonly JniPeerMembers _members_example_MyList = …
    }

This can result in long member names, but we should be well within
C# language limits (512 characters?), and it's readable!