icanhaz blittable method registration?
Commit db84cc3d "didn't work" because the "traditional" form of
method registration through `JNIEnv::RegisterNatives()` used a
`System.Delegate` instance to refer to a C# method to invoke, but
this fails under NativeAOT because the NativeAOT infrastructure needs
to be able to produce "stubs" for the methods based on compile-time
scans, and it doesn't represent our code pattern and/or our pattern
is not supportable with how NativeAOT works.

Oops.

"Ecosystem-wise", I'm not sure how to solve this, other than
handy-wavy ideas such as "rely on `jnimarshalmethod-gen`!".

Short-term, we can at least resolve `ManagedPeer`.

Create a new `JniBlittableNativeMethodRegistration` struct which
contains only blittable types, vs. `JniNativeMethodRegistration`
which contains `string` fields and `Delegate` fields and is not
easily supportable by NativeAOT.

	public partial struct JniBlittableNativeMethodRegistration {
	    public JniBlittableNativeMethodRegistration (IntPtr name, IntPtr signature, IntPtr marshaler);
	    // To support `new JniBlittableNativeMethodRegistration("name"u8, "signature"u8, marshaler)`
	    public JniBlittableNativeMethodRegistration (ReadOnlySpan<byte> name, ReadOnlySpan<byte> signature, IntPtr marshaler);
	}

"Plumb through" use of `JniBlittableNativeMethodRegistration` to
`JniEnvironment.Types.RegisterNatives()` overloads:

	partial class JniEnvironment {
	    partial class Types {
	        public static void RegisterNatives(JniObjectReference type, ReadOnlySpan<JniBlittableNativeMethodRegistration> methods);
	    }
	}

This in turn requires updating `jnienv-gen` so that
`RegisterNatives()` is bound with the `const JNINativeMethod *methods`
parameter bound as an `IntPtr` instead of a
`JniNativeMethodRegistration[]`.

Update the `ManagedPeer` static constructor to use
`[UnmanagedCallersOnly]` methods and native function pointers to
register the marshal methods that `ManagedPeer` requires.