Allow marking constructors as callable from Java.
Continuing on the "activation constructor" train, we want to allow
Java to pass parameters to a constructor.

In .NET Android parlance:

	class Example : Java.Lang.Object {
	    [Export]
	    public Example (int value) {}
	}

Commit 9027591c updated `JavaCallableAttribute` to allow it to be
placed on constructors, but if we try that with
`samples/Hello-NativeAOTFromJNI`, the Java Callable Wrapper
doesn't build:

	// JCW
	/* partial */ class Example extends java.lang.Object {
	    public Example (int p0) {
	        super (p0);
	        if (getClass () == Example.class) ManagedPeer.construct(…);
	    }
	}

To make this work, we need `ExportAttribute.SuperArgumentsString`.

We *could* add this to `JavaCallableAttribute`, but that means
we'd have a property which can only be used from constructors.
(Yes, `ExportAttribute` has this "wonkiness", but that doens't mean
we should continue it!)

Add a new `JavaCallableConstructorAttribute`, which has a new
`SuperConstructorExpression` property, along with `jcw-gen` support.
This allows things to compile:

	// C#
	class Example : Java.Lang.Object {
	    [JavaCallableConstructor (SuperConstructorExpression="")]
	    public Example (int value) {}
	}

	// JCW
	/* partial */ class Example extends java.lang.Object {
	    public Example (int p0) {
	        super ();
	        if (getClass () == Example.class) ManagedPeer.construct(…);
	    }
	}

…but it's not quite right yet, because the wrong constructor is
invoked!

	ManagedPeer.construct (
	    /* self */      this,
	    /* aqn */       "Namespace.Example, Assembly",
	    /* ctor sig */  "",
	    /* arguments */ new java.lang.Object[] { p0 }
	);

Oops!

Update the `Signature`/`SignatureOptions` creation within
`JavaCallableWrapperGenerator.cs` so that `ManagedParameters`
is set for *all* constructor codepaths.

	ManagedPeer.construct (
	    /* self */      this,
	    /* aqn */       "Namespace.Example, Assembly",
	    /* ctor sig */  "System.Int32, System.Runtime",
	    /* arguments */ new java.lang.Object[] { p0 }
	);

This appears to fix a long-standing bug/"thinko" in JCW generation!