Avoid DotNetCoreCLI@2
Through the glory of printing the contents of the directory
after `dotnet publish` and as part of `dotnet build -t:RunJavaSample`,
we see that the `$(PublishDir)` `bin/Release/osx-x64/publish`
exists after `dotnet publish`, but is *gone* when `dotnet build`
runs (?!).  In its place is a `publish.zip` file, which doesn't do
us any good!

We thus infer and assume that the DotNetCoreCLI@2 task is creating
this `publish.zip` file and removing the directory, which prevents
the execution step from actually executing.

Replace DotNetCoreCLI@2 with a powershell script which directly invokes
`dotnet publish` & co.  Hopefully this will avoid the problem.