[Java.Interop.Tools.JavaCallableWrappers] don't build net8.0 in XA
xamarin/xamarin-android's build began failing with:

    (CoreCompile target) ->
    Xamarin.Android.Build.Tasks\Utilities\MamJsonParser.cs(92,43): error CS0122: 'NotNullWhenAttribute' is inaccessible due to its protection level [Xamarin.Android.Build.Tasks\Xamarin.Android.Build.Tasks.csproj]
    Xamarin.Android.Build.Tasks\Utilities\MamJsonParser.cs(92,81): error CS0122: 'NotNullWhenAttribute' is inaccessible due to its protection level [Xamarin.Android.Build.Tasks\Xamarin.Android.Build.Tasks.csproj]
    Xamarin.Android.Build.Tasks\Utilities\MavenExtensions.cs(26,32): error CS0122: 'NotNullWhenAttribute' is inaccessible due to its protection level [Xamarin.Android.Build.Tasks\Xamarin.Android.Build.Tasks.csproj]

This was happening because the `net8.0` build of
`Java.Interop.Tools.JavaCallableWrappers.dll` was being used and
should not be: we want the `netstandard2.0` version to be used for
MSBuild task assemblies. To fix this, let's only build
`netstandard2.0` when `$(XABuild)` is `true`.

We should also avoid using `$(AppendTargetFrameworkToOutputPath)=false`
when multi-targeting: stop using it inside XA's build.