[generator] Extend `skipInvokerMethods` support to interfaces. (#1202)
Context: 73ebad2496a0d6df251fd24e216248da70e3fea9
Context: https://github.com/xamarin/AndroidX/pull/779

Commit 73ebad24 added support for the metadata attribute
`//class[@skipInvokerMethods]`, which allowed us to suppress generation
of invoker methods for a class.

The AndroidX Media3 binding in xamarin/AndroidX#779 hits a similar
issue wherein we generate incorrect generics on an invoker type.
However, this is an invoker type for an *`interface`*, not a `class`.

Extend our `skipInvokerMethods` support so that it can be used as
`//interface[@skipInvokerMethods]` as well.