Use the mcr.microsoft.com/devcontainers/universal:2-linux image
Not only does this project use C# and Java, but it *also* contains
C++ code in `src/java-interop`, which meant the previous attempt
failed to build.

Instead of trying to find the minimal set of dependencies needed to
build, just go with the default universal:2-linux image, which
should have all needed dependencies.