[generator-Tests] bind java.util.List
Context: 0aec86ac404ec6a181e635b73838f108345fed9c

Commit 0aec86ac mentioned:

> Remaining Integration-Tests which don't support JavaInterop1:
>
>  * Missing builtin mapping between collection types:
>
>        warning BG8800: Unknown parameter type 'java.util.List<java.lang.Integer>' for member …
>        warning BG8800: Unknown parameter type 'java.util.ArrayList<java.lang.Integer>' for member
>        warning BG8400: Unexpected field type `java.util.List<java.lang.String>`
>
>    Impacts: Java_Util_List.cs, NormalMethods.cs, ParameterXPath.cs

The problem here is that the provided API XML did not contain a
declaration for `java.util.List<E>`.  This didn't matter for
XAJavaInterop1 output, as it binds `List<E>` as
`System.Collections.Generic.IList<T>`, but it *does* matter for
JavaInterop1 output, as it doesn't special-case `List<E>`.
(Perhaps it should!  Or this helps "lower" the need for improved
generics! Or…)

Add a declaration for `java.util.List<E>` to Java_Util_List.cs,
and ParameterXPath.cs, so that we can set `TryJavaInterop1 => true`.

Also add a declaration for `java.util.List<E>` to NormalMethods.cs.
This doesn't allow us to enable `TryJavaInterop1 => true`, because
of the "`Handle` isn't a public property" issue, but it'll be a
necessary prerequisite to enable NormalMethods.cs in the future.