[Java.Interop] Add JniTypeSignatureAttribute.InvokerType
Context: https://github.com/dotnet/java-interop/issues/1263
Context: https://github.com/dotnet/java-interop/issues/1263#issuecomment-2517292793

We want the default trimmer infrastructure to be able to automatically
preserve the `*Invoker` types which are required for interacting with
`abstract` classes and interfaces.

The most straightforward way to do this is to add a new `InvokerType`
property to `JniTypeSignatureAttribute` (and eventually
`RegisterAttribute`):

	partial class JniTypeSignatureAttribute {
	    [DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicConstructors |
	        DynamicallyAccessedMemberTypes.NonPublicConstructors)]
	    public Type? InvokerType {get; set;}
	}

Update `generator` so that `generator --codegen-target=JavaInterop1`
output sets this new property on abstract classes and interfaces:

	namespace Java.Lang {

	    [Java.Interop.JniTypeSignatureAttribute("java/lang/Runnable", GenerateJavaPeer=false, InvokerType=typeof(Java.Lang.IRunnableInvoker))]
	    public partial interface IRunnable {
	    }
	    internal partial class IRunnableInvoker {
	    }

	    [Java.Interop.JniTypeSignatureAttribute("java/lang/Number", GenerateJavaPeer=false, InvokerType=typeof(Java.Lang.NumberInvoker))]
	    public abstract partial class Number {
	    }
	    internal partial class NumberInvoker {
	    }
	}

This allows the default trimmer to automatically preserve the
`*Invoker` type and constructors.

Update `Java.Interop.JniRuntime.JniValueManager` to no longer look
for `*Invoker` types "by string", and instead require use of the
`JniTypeSignatureAttribute.InvokerType` property.

Update unit tests and expected output so that everything works.

Note: XAJavaInterop1 output is impacted, but it shouldn't matter, as
it changes the `RegisterAttribute` `connector` on interfaces,
which isn't *used* by anything:

	// old and busted
	[Register ("xamarin/test/NotificationCompatBase$Action$Factory", "", "Xamarin.Test.NotificationCompatBase/Action/IFactoryInvoker")]

	// new hawtness
	[Register ("xamarin/test/NotificationCompatBase$Action$Factory", "", "Xamarin.Test.NotificationCompatBase.Action.IFactoryInvoker")]

Further note that the old value is non-sensical.